python_libs_test.py
import cryptography
from cryptography.fernet import Fernet
import PIL
from PIL import Image
import reportlab
from reportlab.lib.pagesizes import letter
import streamlit
import twilio
from twilio.rest import Client
import mercadopago
import fastapi
from fastapi import FastAPI
import uvicorn
import dotenv
from dotenv import load_dotenv
import mysql.connector
import pymongo
from pymongo import MongoClient

print("Todas las librerías importadas correctamente")