requirements.txt
pyngrok==7.2.0
cryptography==43.0.1
pillow==10.4.0
reportlab==4.2.2
streamlit==1.37.1
twilio==9.2.3
mercadopago==2.2.0
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
mysql-connector-python==9.0.0
pymongo==4.8.0